Report missing wikidata from latest applicable term
Rather than only reporting on missing Wikidata identiiers for members of
the most recent term (with an empty report if everyone in that term has
been reconciled), report on everyone unreconciled in the most recent
term that has not yet been fully reconciled.

In other words, if the most recent term has been fully reconciled, step
back to the previous terms, until it in turn is fully reconciled, and so
on.