Belgium (Representatives): choose explicit fields from Wikipedia
We're about to migrate away from this being our primary source, so be
more precise in what we want from it. In particular, remove the "source"
column, so that we isolate diffs to the membership records later.

NB: The upstream is out of sync, so I rebuilt this locally with:
  q -H -d, -O "SELECT name,party,area,term,start_date,end_date FROM wikipedia.csv"