Canada (Commons): explicitly specify Wikipedia columns
The scraper now includes various other columns (eg. reconciling people,
parties, and areas to Wikidata.) These are conceptually useful, but not
yet, as (for example) some of the areas have been redistricted, and the
Wikipedia page for the old name has been redirected to the new name. So
even though there are separate Wikidata items for each the scraper only
gets the ID for the new one.

Once we switch Term 42 to being sourced from Wikidata instead, it'll be
easier to use the IDs for Term 41.