Look up file change times via Octokit if no local history
The everypolitician-bot works with a shallow clone of the repo, with no
history, so if it asks `git` for the last change time of a file, it will
be seconds ago. So make it so that we look up the date via the Github
API instead, unless EP_FULL_GIT is set in the environment.