Use jsDelivr rather than RawGit for CDN
RawGit is shutting down [1], so we need to switch to another CDN provider
that supports GitHub. Luckily jsDelivr [2] does exactly this, so only a
minimal change is required.

This change means that all of the URLs for Popolo JSON and CSV files in
`countries.json` will change next time that file is rebuilt.

[1]: https://rawgit.com/
[2]: https://www.jsdelivr.com/rawgit